NIFI-6684 Add more property to Hive3ConnectionPool
